refactor: encapsulate `create_stream_and_schema_from_storage`
