+++ src/handlers/http/modal/utils/logstream_utils.rs
@@ -32,8 +32,9 @@ use crate::{
     },
     metadata::{self, SchemaVersion, STREAM_INFO},
     option::{Mode, CONFIG},
+    parseable::PARSEABLE,
     static_schema::{convert_static_schema_to_arrow_schema, StaticSchema},
-    storage::{LogStream, ObjectStoreFormat, StreamType},
+    storage::StreamType,
     validator,
 };
 use tracing::error;
@@ -64,7 +65,9 @@ pub async fn create_update_stream(
 
     if !metadata::STREAM_INFO.stream_exists(stream_name)
         && CONFIG.options.mode == Mode::Query
-        && create_stream_and_schema_from_storage(stream_name).await?
+        && PARSEABLE
+            .create_stream_and_schema_from_storage(stream_name)
+            .await?
     {
         return Err(StreamError::Custom {
             msg: format!(
@@ -445,69 +448,6 @@ pub async fn create_stream(
         }
     }
     Ok(())
-}
-
-/// list all streams from storage
-/// if stream exists in storage, create stream and schema from storage
-/// and add it to the memory map
-pub async fn create_stream_and_schema_from_storage(stream_name: &str) -> Result<bool, StreamError> {
-    // Proceed to create log stream if it doesn't exist
-    let storage = CONFIG.storage().get_object_store();
-    let streams = storage.list_streams().await?;
-    if streams.contains(&LogStream {
-        name: stream_name.to_owned(),
-    }) {
-        let mut stream_metadata = ObjectStoreFormat::default();
-        let stream_metadata_bytes = storage.create_stream_from_ingestor(stream_name).await?;
-        if !stream_metadata_bytes.is_empty() {
-            stream_metadata = serde_json::from_slice::<ObjectStoreFormat>(&stream_metadata_bytes)?;
-        }
-
-        let mut schema = Arc::new(Schema::empty());
-        let schema_bytes = storage.create_schema_from_ingestor(stream_name).await?;
-        if !schema_bytes.is_empty() {
-            schema = serde_json::from_slice::<Arc<Schema>>(&schema_bytes)?;
-        }
-
-        let mut static_schema: HashMap<String, Arc<Field>> = HashMap::new();
-
-        for (field_name, field) in schema
-            .fields()
-            .iter()
-            .map(|field| (field.name().to_string(), field.clone()))
-        {
-            static_schema.insert(field_name, field);
-        }
-
-        let time_partition = stream_metadata.time_partition.as_deref().unwrap_or("");
-        let time_partition_limit = stream_metadata
-            .time_partition_limit
-            .and_then(|limit| limit.parse().ok());
-        let custom_partition = stream_metadata.custom_partition.as_deref().unwrap_or("");
-        let static_schema_flag = stream_metadata.static_schema_flag;
-        let stream_type = stream_metadata
-            .stream_type
-            .map(|s| StreamType::from(s.as_str()))
-            .unwrap_or_default();
-        let schema_version = stream_metadata.schema_version;
-        let log_source = stream_metadata.log_source;
-        metadata::STREAM_INFO.add_stream(
-            stream_name.to_string(),
-            stream_metadata.created_at,
-            time_partition.to_string(),
-            time_partition_limit,
-            custom_partition.to_string(),
-            static_schema_flag,
-            static_schema,
-            stream_type,
-            schema_version,
-            log_source,
-        );
-    } else {
-        return Ok(false);
-    }
-
-    Ok(true)
 }
 
 /// Updates the first-event-at in storage and logstream metadata for the specified stream.
