+++ src/lib.rs
@@ -39,6 +39,7 @@ pub mod migration;
 mod oidc;
 pub mod option;
 pub mod otel;
+mod parseable;
 mod query;
 pub mod rbac;
 mod response;
